% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixmodel_S3classfunctions.R
\name{as_socmixmod_fittedmodels}
\alias{as_socmixmod_fittedmodels}
\title{Create object of class as_socmixmod_fittedmodels}
\usage{
as_socmixmod_fittedmodels(summary.table, allmodels.list, criteria)
}
\description{
Create object of class as_socmixmod_fittedmodels
}
