% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixmodel_S3classfunctions.R
\name{as_socmixmod_model}
\alias{as_socmixmod_model}
\title{Create a soc mix model object}
\usage{
as_socmixmod_model(
  K.in,
  K.out,
  Den,
  Num,
  Means,
  Freqs,
  S,
  rho,
  loglik,
  BIC,
  AIC,
  ICL,
  nrep,
  mu
)
}
\value{
The \code{socmixmodel_model} class is a wrapper for the listed output of a
binomial association mixture model. Individual list elements can still be accessed in
the usual way (\code{$} or \code{[[]]}).

Functionality with generic methods \code{summary}
(\code{summary.socmixmod_model}) and plot (\code{print.socmixmod_model}).

Also works with customised generic method \code{by_edge} to output the
probabilities of a each dyad beloning to each component
}
\description{
Creates an S3 object of class \code{socmixmodel_model}. From the outhput of
a binomail association mixture model (\code{fit_binom_mixture_model})
}
\examples{
dat = socmixmod_simulate_data()
mod = fit_binom_mixture_model(Den = dat$d, Num = dat$n, J = dat$K, edge.ids = dat$edge.ids)
summary(mod)
print(mod)
by_edge(mod)
}
