% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixmodel_extras.R
\name{by_edge}
\alias{by_edge}
\title{Extract and tidy the by edge data in a socmixmodel_model class object}
\usage{
by_edge(obj)
}
\arguments{
\item{obj}{a socmixmodel_model class object}
}
\value{
A dataframe showing each edge in the population, the probability
they belong to each component \code{P(ki)} and the component the edge
is most likely to belong to
}
\description{
This function takes the data from a socmixmodel_model class object (produced
with \code{fit_binom_mixture_model}) and produces a table of data by edge. It
also calculates the component a given edge belongs to based on the one which
if has the highest probability of belonging to.
}
\examples{
#simulate data
dat = socmixmod_simulate_data()

# Run model
fitted.model =
fit_binom_mixture_model(Den = dat$d, Num = dat$n, J = dat$K, edge.ids = dat$edge.ids)

#View output
fitted.model
summary(fitted.model)
by_edge(fitted.model)

}
