% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixmodel_extras.R
\name{get_bestmodel}
\alias{get_bestmodel}
\title{Get the best model method from a  socmixmod_fittedmodels object}
\usage{
get_bestmodel(obj, verbose = TRUE)
}
\arguments{
\item{obj}{socmixmod_fittedmodels class object}

\item{verbose}{Logical. If TRUE displays summary information from the best fitting
model when applied. Default=TRUE.}
}
\value{
A socmixmodel_model of the best fitting model as specified by the
\code{criterion} choosen for model fitting.
}
\description{

}
\examples{
# simulate data
dat = socmixmod_simulate_data()

# Run model
model.fitting =
binom_assoc_mixt(Den = dat$d, Num = dat$n, edge.ids = dat$edge.ids, run.all.J = TRUE)

#View output
model.fitting
summary(model.fitting)
best.mod = get_bestmodel(model.fitting)
summary(best.mod)

}
\seealso{
binom_assoc_mixt
}
