% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixmodel_S3classfunctions.R
\name{plot.socmixmod_model}
\alias{plot.socmixmod_model}
\title{Plotting method for socmixmod_model objects}
\usage{
\method{plot}{socmixmod_model}(obj)
}
\arguments{
\item{obj}{a socmixmodel_model object}
}
\value{
A ggplot density plot.
}
\description{
This produces a density plot of edge SRI by component to how the binomial
mixture model maps onto the real data. It is important to note that the
Binomial Mixture Models do not use SRI in their calculation and this plot is
intended simply as a guide.
}
\details{
Plot requires ggplot2.
}
