% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixmodel_extras.R
\name{socmixmod_simulate_data}
\alias{socmixmod_simulate_data}
\title{Simulate Data}
\usage{
socmixmod_simulate_data(K = 3, N = 20, mean.d = 80)
}
\arguments{
\item{K}{Number of components to simulate in the data (default 3)}

\item{N}{Number of individuals to simulate (default = 20)}

\item{mean.d}{Mean sampling effort per dyad (default = 80)}
}
\value{
Output is a list with elements:
#' \itemize{
 \item \code{d}: SRI Denominator. Number of times simulated dyad are
 observed in total.
 \item \code{n}: SRI Numerator. Number of times simulated dyad are observed
 together.
 \item \code{K}: Number of components (will match inputted K)
 \item \code{edge.ids}: simulated edge id names
}
}
\description{
Function to simulate data with a known number of components (K) to use with
socmixmod functions. See Weiss et al 2019 for details.
}
\examples{
socmixmod_simulate_data()

socmixmod_simulate_data(K = 5, N = 100, mean.d=50)

}
