% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixmodel_S3classfunctions.R
\name{summary.socmixmod_fittedmodels}
\alias{summary.socmixmod_fittedmodels}
\title{Summary generic method for socmixmod_fittedmodels}
\usage{
\method{summary}{socmixmod_fittedmodels}(obj)
}
\description{
Summary generic method for socmixmod_fittedmodels
}
