% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixmodel_S3classfunctions.R
\name{summary.socmixmod_model}
\alias{summary.socmixmod_model}
\title{Generic summary for soc mix model object}
\usage{
\method{summary}{socmixmod_model}(obj)
}
\description{
Generic summary for soc mix model object
}
\seealso{
\link[socmixmods]{as_socmixmod_model}
}
