% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixmodel_functions.R
\name{binom_assoc_mixt}
\alias{binom_assoc_mixt}
\title{Fit Binomial mixture models to  social association data}
\usage{
binom_assoc_mixt(
  Den,
  Num,
  edge.ids = NULL,
  minJ = 1,
  maxJ = 9,
  nrep = 20,
  criterion = "ICL",
  run.all.J = FALSE
)
}
\arguments{
\item{Den}{A numeric vector of the association index denominator.}

\item{Num}{A numeric vector of the association index numerator.}

\item{minJ}{Integer. Minimum component model to fit to the data. Default = \code{1}.}

\item{maxJ}{Integer. Maximum component model to fit to the data. Default = \code{9}.}

\item{nrep}{Something}

\item{criterion}{String. Model fitting criteria to use to choose best model.
Categorical, options are \code{"ICL"}[default], \code{"AIC"} & \code{"BIC"}.}

\item{run.all.J}{Logical. If \code{TRUE} fits all component models between \code{minJ}
and \code{maxJ}. If \code{FALSE} [default], when two component-models in a row fit less well (by
\code{criterion}) stops the analysis and returns output.}
}
\value{
Object of class \code{socmixmod_fittedmodels}.Features of the
model can be accessed with the \code{summary} and
\code{get_bestmodel} functions.

In addition, the output can be treated as a list. The output list has two
elements.

\enumerate{
\item \code{summary.table} A matrix of summary information from all fitted
models. See also \code{summary(MODEL)}. Each row is a model
fitted with k components with columns:
 \itemize{
 \item \code{K.in}: Input number of components \code{J}
 \item \code{K.out}: Number of components fitted by the model
 (sanity check, should match \code{K.in}).
 \item \code{S}: Shannon's Entropy.
 \item \code{rho}: Overdispersion parameter
 \item \code{AIC}: Model AIC
 \item \code{BIC}: Model BIC
 \item \code{ICL}: Model ICL
}

\item \code{all.models}. A list of length \code{max(k)}. Each element in the
list is a \code{socmixmod_model} class object outputted from
\code{fit_binom_mixture_model}. List is in ascending order from \code{minJ}.
See also \code{get_bestmodel(OUTPUT)} to extract the best model from the list.
Note this can also be done by simply selecting the index of the best mode
from this list: \code{OUTPUT$all.models[[best.model.index]]}

\item \code{criteria} Criteria inputted by the user used to select the best
model.
}
}
\description{
This function fits up to \code{maxJ} binomial mixture models to social data. Returning a summary
of the fit of each model and the output of the best fitting model. \cr \cr
Where J > 1 this calls function \code{fit_binom_mixture_model}. Where J = 1, fits a standard
(non-mixture) binomial model. If minJ=maxJ a single model is run.
}
\examples{
#simulate data
dat = socmixmod_simulate_data()

# Run model
model.fitting =
binom_assoc_mixt(Den = dat$d, Num = dat$n, edge.ids = dat$edge.ids, run.all.J = TRUE)

#View output
model.fitting
summary(model.fitting)
best.mod = get_bestmodel(model.fitting)
summary(best.mod)
}
