% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixmodel_functions.R
\name{fit_binom_mixture_model}
\alias{fit_binom_mixture_model}
\title{Fit a J component binomial mixture model}
\usage{
fit_binom_mixture_model(
  Den,
  Num,
  J,
  edge.ids = NULL,
  maxiter = 1000,
  maxrep = 50,
  tolfun = 1e-06,
  minprior = 0
)
}
\arguments{
\item{Den}{A numeric vector of the association index denominator.}

\item{Num}{A numeric vector of the association index numerator.}

\item{J}{Integer number of components to fit.}

\item{edge.ids}{String vector of edge names. If \code{NULL} (default) then
names will be created and applied within the function.}

\item{maxiter}{Number of iterations. Default input = 1000.}

\item{maxrep}{Something. Default input = 50}

\item{tolfun}{Somehing else. Default input = 1e-06.}

\item{minprior}{Something else again. Default input = 0.}
}
\value{
Output is an object of the class \code{socmixmod_model}. Features of the
model can be accessed with the \code{summary}, \code{plot}, and
\code{by_edge} functions.

All model output can additionally be accessed by treating the output as a list
(i.e. via \code{$} or \code{[[]]}). The output list contains the following
elements:

\itemize{
 \item \code{K.in} Integer of components aimed to be fitted
 (will match input \code{J}).
 \item \code{K.out} Integer number of components fitted by the model. Returned as a
 sanity check. Should match \code{K.in} & \code{J}.
 \item{Denominators}. numeric vector \code{den} as Inputted.
 \item{Numerators} Numeric vector \code{Num} as inputted
 \item \code{Mean} Numeric vector of the mean association rates of the \code{J}
 fitted components.
 \item \code{Frequency} Numeric vector indicating the proportion of
 associations in the population belonging to each component.
 \item \code{S} Shannons' Entropy (social complexity measure)
 \item \code{rho} Numeric realised overdispersion parameter.
 \item \code{logLik} Model log-likelhood
 \item \code{BIC} Model Baysien Information Criteria
 \item \code{AIC} Model Akalie Information Criteria.
}
}
\description{
\code{fit_binom_mixture_model} fits a J component mixture model to social association
data. This function will usually be only called through \code{\link{binom_assoc_mixt}} but it
can be used as a stand-alone function.
}
\examples{
#simulate data
dat = socmixmod_simulate_data()

# Run model
fitted.model =
fit_binom_mixture_model(Den = dat$d, Num = dat$n, J = dat$K, edge.ids = dat$edge.ids)

#View output
fitted.model
summary(fitted.model)
by_edge(fitted.model)

}
\seealso{
\code{\link{binom_assoc_mixt}}
}
